多行删除

+++ user_case/src/com/neusoft/service/impl/UserServiceImpl.java
@@ -49,9 +49,15 @@ public class UserServiceImpl implements UserService {
         return dao.findUserByUsernameAndPassword(user.getUsername(),user.getPassword());
     }
 
+
     @Override
-    public void delSelectUser(String id) {
-        dao.delete(Integer.parseInt(id));
+    public void delSelectUser(String[] ids) {
+        // 根据id批量删除数据
+        if (ids!=null && ids.length>0){
+            for (String id:ids) {
+                dao.delete(Integer.parseInt(id));
+            }
+        }
     }
 
     @Override
